Removes head_tag_content override to pick up Google Scholar tags from Sufia. Fixes #620.
This work required first getting ScholarSphere working on top of Sufia 4.1.0, which included proxy- and transfer-related code ported from ScholarSphere, so this meant a lot of deletion and cleanup. Then I needed to tweak Sufia to allow downstream head content injection, which was merged and is now pinned in our Gemfile. The last step was creating a new override in ScholarSphere to inject the favicon_link_tag that isn't in Sufia. A new spec test is included to make sure the favicon exists.