Removes a lot of old images, creates a service for the collection thumbnail, adds new collection png. Adds indexer and refs image path, fixes rubocop errors.
Collection indexer inherits from CC Indexer, namespace Sufia for better compatibility, updates collection.rb for new collection indexer.

Tests for the presence of the correct collection PNG and not the default image.

Trying the have_css matcher to test the presence of the correct png file.

Fixes the test by using ensuring that the thumbnail path starts with assets/collection