Navbar with search is consistent with other pages. Overrides hompage index from Sufia to start customizing page. Hides other partials to work on custimization. Moves the share section out of the index, removes a lot of extra divs and styles, makes a row and two columns for the marketing text and share.
Adds typography style sheet to override default Bootstrap styles. Fixes up some styles with overrides and roughs in the basic spacing and layout.

Adds the announcment box back in, super horrible overrides for style, and adds moar bootstrap syntax. Updates the styles for the announcement section and uses moar ugly overrides because of the very specific declarations from Sufia. Adjusts spacing and styles for the announcement section and the home tabs with content.

Moves the search box to its own row below the marketing row. Removes the white background on the announcment text and makes the text white. Fixes contrast errors.