Overrides construct_page_title for vertical bar and defines page_title in work show presenter.
Adds the collection show page so that we can overide the page title to keep it consistent with the style of the work show page.

Updates edit page for collections to follow style from work edit page.

Removes the construct_page_title method to just pass a string with dynamic elements to output contents of title element.