Rename buttons in create collection page, ref #590
When creating a new collection, three different buttons will appear:
1) creating a new, empty collection
2) creating a new collection, then redirecting to the batch create page with the newly created collection pre-selected
so that you may add new works to the collection.
3) creating a new collection, then redirecting to the dashboard works page so that you can select existing works to add
to the new collection.

Additionally, when doing the reverse of option 3 above, and selecting works from the dashboard first, then creating a new
collection, only one button is shown. Also, a listing of the works that will be added to the collection is displayed
using the same UI components that are used when displaying works in existing collections.