Making the single-use links more accessible, ref #738
Clicking the button to create a new single-use link wasn't making an obvious change.

The button was moved down under the table to which the links were being added, enabling the user to see the result more easily.
Additionally, aria labels and more informative text was added so that visually impaired links could more readily understand
what was occurring and what the links and buttons do.

The actions still aren't ideal. A better solution will be to not use tables and update the Javascript to include responses with
better accessibility functions.