Provide separate step for uploading all files, ref #453
Auto-uploading multiple small files was intermittently failing.

This solution injects a new UI behavior where auto-uploading is disabled and a new button is provided to
being uploading all the files after their records are individually added to the DOM.

Additionally, the JS templates and other UI code are amended to provide better accessibility. This includes adding aria
labels, appropriate table headers, and captions so screen readers can interpret the changes made to the table listing
of files.

Although this does add an additional step to the workflow, it fixes the problem and also enables non-sequential,
concurrent uploading for increased performance.