Override LocalFileJob to set file set's title to filename, ref #876
When uploading new versions of files, the file set's title retained the name of the first version of the file. If another version
of the file was uploaded using a different name, the file set's title wasn't updated.

Overriding CurationConcerns LocalFileJob to update the file set title with the file name fixes this issue. Additionally,
the :after_revert_content callback is used to ensure the title and label are changed if the file set is reverted
to a previous version.