Proxied user is the depositor in BaseActor, #895
This is a reattempt of the solution that was first tried (unsuccessfully) in f5cadae53e887df9292afabe4ca950e87412ae6b.

Due to a race condition between GenericWorkActor, AttachFilesToWorkJob and ContentDepositorChangeEventJob, the file set's
depositor could be set at the proxy user instead of the proxied user, that is, the user whom is being deposited on
behalf of.

This addresses the problem by ensuring the work is set to the proxied user at the first opportunity. ContentDepositorChangeEventJob
will still run, which also sets the depositor to the proxied user, but doing it in the actor first ensure that when
AttachFilesToWorkJob runs, it will always have the correct depositor set.

It also refactors the actor code slightly to account for future changes that will need to occur when #948 and #949 are closed.