Adds the proxied user as the depositor in BaseActor, #895
Due to a race condition between GenericWorkActor, AttachFilesToWorkJob and ContentDepositorChangeEventJob, the file set's
depositor could be set at the proxy user instead of the proxied user, that is, the user whom is being deposited on
behalf of.

This addresses the problem by ensuring the work is set to the proxied user at the first opportunity. ContentDepositorChangeEventJob
will still run, which also sets the depositor to the proxied user, but doing it in the actor first ensure that when
AttachFilesToWorkJob runs, it will always have the correct depositor set.