Display NullPresenter when representative is restricted, ref #892
In specific cases when a public work has multiple file sets and its representative is set to non-public one,
the representative presenter will be set to nil, resulting in errors in the show page.

Changing up the logic for determining a NullPresenter fixes this problem. The result is that if the public is viewing a public
work with a restricted representative, the public will see no representative or thumbnail for the work, but will still see
any file sets that are public.

Additionally, this combines a thumbnail feature test into a more comprehensive feature test for thumbnails and representatives
that also runs faster.