Rename file sets during upload errors, ref #974 & #969
ImportUrlJob will fail if the file's url has expired, or if there is
some other kind of error during its retrieval, such as a timeout or
other disruption.

Instead of letting the job fail and having Resque capture the error, we
can have the job capture the error and rename the file set to the kind
of error.

The benefit is that it surfaces the error to the user so that they can
take action: delete the file set and try again. The drawback is that
it's still creating the resource in Fedora. Ideally, perhaps, we ought
to capture these errors even before the file set is created and attached
to the work. However, that would require deeper changes to the stack.