Use a Person model for a work's creator instead of a String. Part of story
For the creator of a work (or collection), we changed the creator to be a Person model instead of a String. The Person model is just a standard ActiveFedora::Base record.

The Person model has properties for first_name and last_name. In the future, there are plans to add properties for ORCID ID, Penn State login, and a list of aliases for different display names.

The work and collection models used to include ::CurationConcerns::BasicMetadata, but now they include ::BasicMetadata instead. Since the creator property was defined in ::CurationConcerns::BasicMetadata, we can no longer include that file because we need to define that property differently. As far as I know, there isn't a way to undefine a property on a model. So, we copied the file over from curation_concerns, and removed the old creator property so that we could re-define it.

On the new/edit form for a work (or collection), adding/removing a creator will now add/remove a set of fields for the associated Person record. The user has the option to try to find an existing Person record (and the auto-suggest will fill in the creator fields), or they can enter new data (and a new Person record will be created when the work is saved).

For the views, presenters, and solr indexing: The solr index keys have changed. The keys are now called creator_name_* because the old keys, creator_* are now used to store the relationship to the Person record. So now creator_name_ssim will index the Person's first and last name, and creator_sim will index the URI to the Person record.

There is a PersonsController that the javascript uses to query Person names for the type-ahead auto-suggest.

There was some import/export code, mostly under lib/scholarsphere/import and some associated files under spec/fixtures. We decided not to spend the effort to update that code for the new Person records because that importer isn't being used anymore. So, we deleted all that code.