Updates paginate to use the member presenters and to correct the route for the main app.
Updates controller to use file_page instead of page because page is being used by something else.

Updates spec test to check for 10 files on page 1 and look for a link for the second page.

Updates test to make sure there are records with data on page 23.

Adds comment to paginate partial and refactors test for the correct range of files.