Pass display_name from form to model for updating
When a display name is edited or created in a work's form, it needs to
be passed to the resource model so it can be persisted.

CreatorForm is used as an intermediate, nested form to hold the display
name, given, and surname values from the Alias and Person objects.
Because "creator name" is still used for display in show views, a
separate display name value is no longer needed. Adjustments were also
made to batch edit so that the form displays all the creator values for
a give batch of works.