Batch edit: setting embargo and lease, ref #1045
Setting an embargo or lease via batch edit would fail.

This enables the setting/overriding of visibilities, embargoes, and
leases in batch edit mode by relying on the actor stack to perform the
same actions on each work in the batch, as would normally be done if the
work was updated alone.

When the works have existing embargoes or leases, and a different one is
supplied, the newer simply overwrites the existing one. If the
visibility is change, the embargo or lease is expired. This may result
in a slight amount of information loss because the history message will
use the embargoed or leased visibility, and not the actual visibility
value specified in the batch update.