refactors the permissions tab for fiieldsets to more closely resemble the pemissions, visibility, sharing section of the work form. Addresses accessibility concerns for the ticket, but not overall for the page. Additional tickets need to be created for the bootlint and WAVE errors.
Adds feature test to ensure that the accessibility work to fix the readability and document outline remains intact.

Reworks the partials to just use other partials and not hard code content in a partial.

Fixes the order of text and quotes in the shaare with partial, adds account name to locale file, removes formgroup and row from form permission partial, and adds form share partial from base with a div with the data-param-key.

Outputs the object name in the share form.

Adds if statement to the headers to look for oject name and output text.

Updates test to check for text file instead of text work.

Refactors if else statement to use a local variable for the object name and pass the translation into the header for the sharing section of the works and files.