Add files to fedora as external files (#1271)
* Add files to fedora as external files

Test for fedora re-direct -- if they have been
added correctly, Fedora should re-direct the
request.

Redefine the predicate for `mime_type` on the
`MetadataNode` class in the Samvera stack so
it does not conflict with the Fedora mime_type,
which is used to handle external content.

Encapsulate behavior behind the environment variable
`ENV['REPOSITORY_EXTERNAL_FILES']`. If that is set
to true the new behavior will be used. If it isn't set
to true or is missing, the old behavior will be used.

If it is true the `ENV['REPOSITORY_FILESTORE_HOST']` and
`ENV['REPOSITORY_FILESTORE']` environment variables must also
be set. See the `rails_helper` for an example of how
to set the variables.

This also overrides `ActiveFedora::File`, `Hydra::Core::DownloadBehaviour`
and `CurationConcerns::WorkingDirectory` to get download functionality
working correctly.

* Add a class to do en masse conversion to external files storage

* Authenticate to Fedora and redirect from http to https