Add a rake task to run thge ExternalFilesConversion class
This commit introduces a rake task that creates a new versions of
a file in order to create a copy of the file that exists on external
storage.

To run the rake task:

`rake scholarsphere:internal_to_external_files`

It will iterate over all the `GenericWork`s in the repository.