Add additional checksum verification in specs
This adds an additional checksum verification to
the external files conversion spec. This ensures
that the FixityService returns the sha1 we expect
for the world.png file.

It also ensures that the checksums are the same
when the existing file is placed in `/tmp` and
that the file that is downloaded from a redirect
also has the correct checksum.