NetworkIngestService to add files to works
Using a configured network ingest directory, existing works are located
based on the directory name in the ingest directory, and any contained
files are added to the work using the existing actor/job process found
in GUI uploads.

This also instals the resque-cleaner gem to provide an improved
interface in the resque admin GUI to retry failed jobs based on their
error messages.