Create zip files for large works and collections
Defines a rake task that creates zips of public collections and works
that exceed a given size. A second rake task will delete zip files
that no longer fit the criteria for the export.

The option to download works or collections in the UI will be presented
if the resource is below the certain threshold size, or if the pre-made
zip is available. If the resource exceeds the specified size and a zip
file has not been created yet, then no zip download option will be
present.